+++ cdap-app-templates/cdap-etl/cdap-data-streams/src/test/java/co/cask/cdap/datastreams/preview/PreviewDataStreamsTest.java
@@ -18,7 +18,11 @@ package co.cask.cdap.datastreams.preview;
 
 import co.cask.cdap.api.data.format.StructuredRecord;
 import co.cask.cdap.api.data.schema.Schema;
+import co.cask.cdap.api.dataset.lib.cube.AggregationFunction;
+import co.cask.cdap.api.dataset.lib.cube.TimeValue;
 import co.cask.cdap.api.dataset.table.Table;
+import co.cask.cdap.api.metrics.MetricDataQuery;
+import co.cask.cdap.api.metrics.MetricTimeSeries;
 import co.cask.cdap.app.preview.PreviewManager;
 import co.cask.cdap.app.preview.PreviewRunner;
 import co.cask.cdap.app.preview.PreviewStatus;
@@ -36,10 +40,13 @@ import co.cask.cdap.proto.ProgramType;
 import co.cask.cdap.proto.artifact.AppRequest;
 import co.cask.cdap.proto.artifact.ArtifactSummary;
 import co.cask.cdap.proto.artifact.preview.PreviewConfig;
+import co.cask.cdap.proto.id.ApplicationId;
 import co.cask.cdap.proto.id.ArtifactId;
 import co.cask.cdap.proto.id.NamespaceId;
 import co.cask.cdap.test.DataSetManager;
 import co.cask.cdap.test.TestConfiguration;
+import com.google.common.base.Stopwatch;
+import com.google.common.collect.ImmutableMap;
 import com.google.gson.JsonElement;
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -47,11 +54,13 @@ import org.junit.ClassRule;
 import org.junit.Test;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 /**
  * Test for preview for data streams app
@@ -87,8 +96,10 @@ public class PreviewDataStreamsTest extends HydratorTestBase {
     List<StructuredRecord> records = new ArrayList<>();
     StructuredRecord recordSamuel = StructuredRecord.builder(schema).set("name", "samuel").build();
     StructuredRecord recordBob = StructuredRecord.builder(schema).set("name", "bob").build();
+    StructuredRecord recordTest = StructuredRecord.builder(schema).set("name", "test").build();
     records.add(recordSamuel);
     records.add(recordBob);
+    records.add(recordTest);
 
     /*
      * source --> transform -> sink
@@ -110,20 +121,21 @@ public class PreviewDataStreamsTest extends HydratorTestBase {
     AppRequest<DataStreamsConfig> appRequest = new AppRequest<>(APP_ARTIFACT, etlConfig, previewConfig);
 
     // Start the preview and get the corresponding PreviewRunner.
-    final PreviewRunner previewRunner = previewManager.getRunner(previewManager.start(NamespaceId.DEFAULT, appRequest));
+    ApplicationId previewId = previewManager.start(NamespaceId.DEFAULT, appRequest);
+    final PreviewRunner previewRunner = previewManager.getRunner(previewId);
 
     // Wait for the preview to be running and wait until the records are processed in the sink.
     Tasks.waitFor(true, new Callable<Boolean>() {
       @Override
       public Boolean call() throws Exception {
         Map<String, List<JsonElement>> data = previewRunner.getData("sink");
-        return data != null && data.get(DATA_TRACER_PROPERTY) != null && data.get(DATA_TRACER_PROPERTY).size() == 2;
+        return data != null && data.get(DATA_TRACER_PROPERTY) != null && data.get(DATA_TRACER_PROPERTY).size() == 3;
       }
     }, 1, TimeUnit.MINUTES);
 
     // check data in source and transform
-    checkPreviewStore(previewRunner, "source", 2);
-    checkPreviewStore(previewRunner, "transform", 2);
+    checkPreviewStore(previewRunner, "source", 3);
+    checkPreviewStore(previewRunner, "transform", 3);
 
     // Wait for the pipeline to be shutdown by timer.
     TimeUnit.MINUTES.sleep(1);
@@ -133,6 +145,13 @@ public class PreviewDataStreamsTest extends HydratorTestBase {
         return previewRunner.getStatus().getStatus();
       }
     }, 1, TimeUnit.MINUTES);
+
+    // Validate the metrics for preview
+    validateMetric(3, previewId, "source.records.out", previewRunner);
+    validateMetric(3, previewId, "transform.records.in", previewRunner);
+    validateMetric(3, previewId, "transform.records.out", previewRunner);
+    validateMetric(3, previewId, "sink.records.in", previewRunner);
+    validateMetric(3, previewId, "sink.records.out", previewRunner);
 
     // Check the sink table is not created in the real space.
     DataSetManager<Table> sinkManager = getDataset(sinkTableName);
@@ -145,4 +164,36 @@ public class PreviewDataStreamsTest extends HydratorTestBase {
     Assert.assertEquals(expectedNumber, result.get(DATA_TRACER_PROPERTY).size());
   }
 
+  private void validateMetric(long expected, ApplicationId previewId,
+                              String metric, PreviewRunner runner) throws TimeoutException, InterruptedException {
+    Map<String, String> tags = ImmutableMap.of(Constants.Metrics.Tag.NAMESPACE, previewId.getNamespace(),
+                                               Constants.Metrics.Tag.APP, previewId.getEntityName(),
+                                               Constants.Metrics.Tag.SPARK, DataStreamsSparkLauncher.NAME);
+    String metricName = "user." + metric;
+    long value = getTotalMetric(tags, metricName, runner);
+
+    // Min sleep time is 10ms, max sleep time is 1 seconds
+    long sleepMillis = TimeUnit.SECONDS.toMillis(1);
+    Stopwatch stopwatch = new Stopwatch().start();
+    while (value < expected && stopwatch.elapsedTime(TimeUnit.SECONDS) < 5) {
+      TimeUnit.MILLISECONDS.sleep(sleepMillis);
+      value = getTotalMetric(tags, metricName, runner);
+    }
+    // wait for won't throw an exception if the metric count is greater than expected
+    Assert.assertEquals(expected, value);
+  }
+
+  private long getTotalMetric(Map<String, String> tags, String metricName, PreviewRunner runner) {
+    MetricDataQuery query = new MetricDataQuery(0, 0, Integer.MAX_VALUE, metricName, AggregationFunction.SUM,
+                                                tags, new ArrayList<String>());
+    Collection<MetricTimeSeries> result = runner.getMetricsQueryHelper().getMetricStore().query(query);
+    if (result.isEmpty()) {
+      return 0;
+    }
+    List<TimeValue> timeValues = result.iterator().next().getTimeValues();
+    if (timeValues.isEmpty()) {
+      return 0;
+    }
+    return timeValues.get(0).getValue();
+  }
 }
