+++ cdap-ui/app/cdap/components/CardActionFeedback/CardActionFeedback.scss
@@ -79,7 +79,7 @@
     height: calc(100% - 35px);
     overflow: auto;
     resize: vertical;
-    max-height: 17vh;
+    max-height: 14vh;
     word-break: break-all;
 
     pre {
