+++ cdap-ui/app/cdap/components/WizardModal/WizardModal.scss
@@ -16,7 +16,7 @@
 
 @import '../../styles/variables.scss';
 $wizard-header-color: #464a57;
-$wizard-modal-height: 85vh;
+$wizard-modal-height: 82vh;
 $wizard-modal-width: 60vw;
 $wizard-modal-margin-top: 12.5vh;
 $wizard-modal-body-max-height: 360px;
@@ -36,7 +36,7 @@ $wizard-modal-max-width: 1050px;
       border: 0;
       overflow: hidden;
       background: transparent;
-      padding-bottom: 35px;
+      padding-bottom: 70px;
 
       .modal-header {
         h4.modal-title > span {
