+++ cdap-ui/app/directives/log-viewer-preview/log-viewer-preview.less
@@ -30,15 +30,15 @@
               }
             }
             .logs-table {
-              height: ~"calc(100vh - 192px)";
-              height: ~"-moz-calc(100vh - 192px)";
-              height: ~"-webkit-calc(100vh - 192px)";
+              height: ~"calc(100vh - 256px)";
+              height: ~"-moz-calc(100vh - 256px)";
+              height: ~"-webkit-calc(100vh - 256px)";
 
               tbody {
                 /* 46px: height of logs table headers */
-                height: ~"calc(100vh - 192px - 46px)";
-                height: ~"-moz-calc(100vh - 192px - 46px)";
-                height: ~"-webkit-calc(100vh - 192px - 46px)";
+                height: ~"calc(100vh - 256px - 46px)";
+                height: ~"-moz-calc(100vh - 256px - 46px)";
+                height: ~"-webkit-calc(100vh - 256px - 46px)";
               }
             }
           }
