+++ cdap-ui/app/directives/log-viewer/log-viewer.less
@@ -483,12 +483,6 @@ my-log-viewer-preview {
           white-space: normal;
           background-color: @selected-row-color;
 
-          .display-time,
-          .level {
-            position: relative;
-            top: -9px;
-          }
-
           td,
           span {
             white-space: normal;
@@ -519,8 +513,6 @@ my-log-viewer-preview {
           }
           .fa-chevron-down {
             color: #393939;
-            position: relative;
-            top: -12px;
             font-size: 8px;
             bottom: 1px;
             margin-left: 4px;
