+++ cdap-ui/app/directives/timeline/timeline.less
@@ -49,7 +49,8 @@ body {
   }
 }
 
-my-timeline {
+my-timeline,
+my-timeline-preview {
   .timeline-container {
     color: white;
