+++ cdap-ui/app/hydrator/controllers/create/toppanel-ctrl.js
@@ -49,7 +49,7 @@ class HydratorPlusPlusTopPanelCtrl {
     this.setState();
     this.HydratorPlusPlusConfigStore.registerOnChangeListener(this.setState.bind(this));
     this.focusTimeout = null;
-    this.timeoutInMinutes = 15;
+    this.timeoutInMinutes = 1;
 
     if ($stateParams.isClone) {
       this.openMetadata();
@@ -356,6 +356,13 @@ class HydratorPlusPlusTopPanelCtrl {
           this.myAlertOnValium.show({
             type: 'success',
             content: `${pipelinePreviewPlaceholder} has completed successfully.`
+          });
+        } else if (res.status === 'KILLED_BY_TIMER') {
+          // TODO: Remove this when we allow users to specify the timeout
+          let minute = this.timeoutInMinutes <= 1 ? 'minute' : 'minutes';
+          this.myAlertOnValium.show({
+            type: 'success',
+            content: `${pipelinePreviewPlaceholder} has completed successfully after ${this.timeoutInMinutes} ${minute}.`
           });
         } else if (res.status === 'STOPPED' || res.status === 'KILLED') {
           this.myAlertOnValium.show({
