+++ cdap-ui/app/hydrator/toppanel.less
@@ -750,5 +750,14 @@ body.theme-cdap {
       }
     }
     .clearfix { padding: 10px; }
+
+    .preview-start-section {
+      display: flex;
+      align-items: center;
+
+      .preview-start-button {
+        margin-left: auto;
+      }
+    }
   }
 }
